ChatbotHttps/django_chatbot/requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
certifi==2023.7.22
channels==4.0.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.3
colorama==0.4.6
Django==4.1.7
django-ckeditor==6.5.1
django-constance==3.1.0
django-debug-toolbar==4.2.0
django-extensions==3.2.1
django-js-asset==2.0.0
django-picklefield==3.1
django-recaptcha==3.0.0
django-sslserver==0.22
djangorestframework==3.14.0
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib-metadata==6.8.0
joblib==1.2.0
Markdown==3.4.3
pydub==0.25.1
pytz==2022.7.1
regex==2023.3.23
requests==2.31.0
rfc3986==1.5.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.3
tqdm==4.65.0
tzdata==2023.2
urllib3==2.0.7
zipp==3.17.0
